ai/crewai/requirements.txt
crewai>=0.28.0
python-pptx>=0.6.21
python-dotenv>=1.0.0
langchain>=0.0.335
langchain-openai>=0.0.2
matplotlib>=3.7.1
pandas>=2.0.0
numpy>=1.24.0
black>=23.12.0
isort>=5.12.0
